+++ components/esp_http_client/lib/include/http_utils.h
@@ -64,7 +64,7 @@ char *http_utils_get_string_between(const char *str, const char *begin, const ch
  * - New string pointer
  * - NULL: Invalid input
  */
-char *http_utils_join_string(const char *first_str, int len_first, const char *second_str, int len_second);
+char *http_utils_join_string(const char *first_str, size_t len_first, const char *second_str, size_t len_second);
 
 /**
  * @brief      Check if ``str`` is start with ``start``
