+++ components/freertos/port/xtensa/port.c
@@ -457,14 +457,14 @@ void __attribute__((optimize("-O3"))) vPortExitCritical(portMUX_TYPE *mux)
 {
 	vPortCPUReleaseMutex( mux );
 	BaseType_t coreID = xPortGetCoreID();
-	BaseType_t nesting =  port_uxCriticalNesting[coreID];
+	BaseType_t nesting = port_uxCriticalNesting[coreID];
 
-	if(nesting > 0U)
+	if(nesting > 0)
 	{
 		nesting--;
 		port_uxCriticalNesting[coreID] = nesting;
 
-		if( nesting == 0U )
+		if( nesting == 0 )
 		{
 			portEXIT_CRITICAL_NESTED(port_uxOldInterruptState[coreID]);
 		}
@@ -480,7 +480,7 @@ void  __attribute__((weak)) vApplicationStackOverflowHook( TaskHandle_t xTask, c
 	char buf[sizeof(ERR_STR1) + CONFIG_FREERTOS_MAX_TASK_NAME_LEN + sizeof(ERR_STR2) + 1 /* null char */] = { 0 };
 
 	char *dest = buf;
-	for (int i = 0 ; i < sizeof(str)/ sizeof(str[0]); i++) {
+	for (size_t i = 0 ; i < sizeof(str)/ sizeof(str[0]); i++) {
 		dest = strcat(dest, str[i]);
 	}
 	esp_system_abort(buf);
